fix: refactor faq backend logic, modify css, and add transition to help center panel
